User and Tokens tables
Upon logging in on the app homepage, user Google ID and tokens are sent to Django. Django puts these into User and Tokens records; Tokens has FK to User, they have a one-to-one relationship. This is a temporary solution, maybe User and Tokens should be saved in the database after some popup asking the user to give permission for the app to do so, rather than doing this upon logging in.