Jobs have a JobType to distinguish if they log temps or set temps, more fields on JobLog and Job, logging jobs take into account all thermo modes, UI section for setting jobs
If the job is a logging job, SettingTemp will be null, otherwise (its a setting job) it will hold the temp it sets the thermo to on an interval. Descriptions aren't set by the user and generate on the backend based on what vars were passed for the unit of time and the increment number. The 2 JobTypes, Logging and Setting, are seeded