Option to download logging job data as JSON file, new page to recieve JSON file of job data from user and plot it
Click on profile to go to new page (need a better place to put the navigation). Sort of a bandaid for the fact that jobs only last for a week since the app isnt verified and tokens expire faster. If a job is dead, download its data before deleting it. Even when the app is verified and Nest tokens last longer, this feature should still stay