fix for the logging job logic
The way Google writes their JSON gives ECO mode a priority over other modes. Lemme know if u want an explanation of what this entails. The functions in front end that extract mode/temps need to adjusted in the same way where the case for eco mode is evaluated first.